Фиксы в логике, реализую вывод графа социальных объектов с рейтингом выше заданного
