+++ src/main/java/com/team3/CreaterOfSocialGraph/service/CohortCounter.java
@@ -2,6 +2,7 @@ package com.team3.CreaterOfSocialGraph.service;
 
 
 import com.team3.CreaterOfSocialGraph.domain.Graph.SocialGraph;
+import com.team3.CreaterOfSocialGraph.domain.RequestMessage;
 import com.team3.CreaterOfSocialGraph.domain.SocialObject;
 
 import java.util.LinkedList;
@@ -12,27 +13,23 @@ public class CohortCounter {
 
     private Map<Long, SocialObject> vertexList;
 
-
-
-    public SocialGraph getSocialObjectListWithRating(SocialGraph socialGraph) {
+    public SocialGraph getSocialObjectListWithRating(SocialGraph socialGraph, RequestMessage requestMessage) {
 
         vertexList = socialGraph.getAdjVertices();
-
-        // Продолжить отсюда
 
         for (Long key : vertexList.keySet()) {
 
            if (vertexList.get(key).getRating() == 100) continue;
            else {
-
-  //             vertexList.get(key).setRating(getRating(vertexList.get(key).getFriendsList()));
+                vertexList.get(key).setRating(getRating(vertexList.get(key).getFriendsList(), requestMessage));
            }
         }
 
         return socialGraph;
     }
 
-    private int getRating(List<Long> friendsList) {
+    private int getRating(List<Long> friendsList, RequestMessage requestMessage) {
+
         int rating = 0;
 
         int count = 0;
@@ -43,10 +40,31 @@ public class CohortCounter {
                 }
         }
 
-        // Пока что для теста
-        if (count > 10 )
-             rating =  80;
-            else rating = 25;
+        switch (requestMessage.getAttributeName()) {
+            case "city":
+                if (count > 20 ) {
+                    rating =  80;
+                    }
+                else if (count > 10 && count < 20) {
+                    rating =  40;
+                } else rating = 10;
+
+                break;
+            case "university":
+                if (count > 15 ) {
+                    rating =  80;
+                } if (count > 10 && count < 15) {
+                rating =  40;
+                } else rating = 10;
+                break;
+            case "school":
+                if (count > 15 ) {
+                    rating =  80;
+                } if (count > 10 && count < 15) {
+                    rating =  40;
+                } else rating = 10;
+                break;
+        }
 
         return rating;
     }
@@ -57,8 +75,11 @@ public class CohortCounter {
 
         List<SocialObject> listSO = new LinkedList<>();
 
+        Long newId = 0L;
         for (Long key : list.keySet()) {
+            list.get(key).setInId(newId);
             listSO.add(list.get(key));
+            newId++;
         }
 
         return listSO;
