--- src/main/resources/static/js/main.css
@@ -1,3 +1,0 @@
-
-
-@import "static/css/request.css";
